+++ Sources/AttributesBuilder.swift
@@ -5,4 +5,4 @@ public struct AttributesBuilder: AttributesContainer {
     public internal(set) var attributes: [NSAttributedStringKey : Any]
 }
 
-extension AttributesBuilder: _AttributesContainer {}
+extension AttributesBuilder: AttributesMutableContainer {}
